refactor getconformancedeclarationprovider
